@@ -80,37 +83,38 @@
     "jest": "^29.7.0",
     "jest-config": "^29.7.0",
     "jest-mock-axios": "^4.7.3",
-    "lint-staged": "^15.2.7",
+    "lint-staged": "^15.2.10",
     "lodash.camelcase": "^4.3.0",
     "npm-run-all": "^4.1.5",
     "open-cli": "^8.0.0",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "replace-in-file": "^8.1.0",
-    "rimraf": "^5.0.9",
-    "rollup": "2.74.1",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.21.3",
     "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-dts": "^4.2.3",
+    "rollup-plugin-dts": "^6.1.1",
+    "rollup-plugin-esbuild": "^6.1.1",
     "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-license": "^2.9.1",
+    "rollup-plugin-license": "^3.5.3",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-sourcemaps": "^0.6.3",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.36.0",
     "run-s": "^0.0.0",
-    "semantic-release": "^24.0.0",
+    "semantic-release": "^24.1.2",
     "shelljs": "^0.8.5",
     "standard-version": "^9.5.0",
-    "ts-jest": "^29.2.0",
+    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^9.0.0",
-    "typedoc": "^0.26.3",
+    "typedoc": "^0.26.8",
     "typedoc-github-theme": "^0.1.2",
     "typedoc-plugin-extras": "^3.1.0",
     "typedoc-plugin-missing-exports": "^3.0.0",
     "typedoc-plugin-rename-defaults": "^0.7.1",
-    "typescript": "^4.6.3"
+    "typescript": "5.5.4"
   },
   "lint-staged": {
     "{src,test}/**/*.ts": [
